requirements.txt
google-adk
google-generativeai
requests
moviepy
python-dotenv
manim
gtts
easygoogletranslate
cloudscraper 